+++ src/main/java/com/ecom/accounting/Accounting/order365/CompleteOrderService.java
@@ -2,6 +2,7 @@ package com.ecom.accounting.Accounting.order365;
 
 import com.ecom.accounting.entities.CompleteOrderData;
 import com.ecom.accounting.entities.OrderStatus;
+import com.ecom.accounting.repositories.CompleteOrderDataRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,11 +20,12 @@ public class CompleteOrderService {
         return completeOrderDataRepository.findBySellerIdAndOrderStatusAndBankSettlementIsNotNull(sellerId, pageable, orderStatus);
     }
 
-    public Long getOrdersCount(String sellerId, Long month, Long year) {
+
+    public Long getOrdersCount(String sellerId, int month, int year) {
         return completeOrderDataRepository.getOrderCountByMonthAndYear(sellerId,month,year);
     }
 
-    public Long getOrdersCount(String sellerId, Long year) {
+    public Long getOrdersCount(String sellerId, int year) {
         return completeOrderDataRepository.getOrderCountByYear(sellerId,year);
     }
 }
